SKETCH - rearranged file structure yet again. Renamed classes
