+++ mchartanalyzer/logger.py
@@ -8,7 +8,7 @@ class Logger:
     def __init__(self, testMode=None):
         self.moduleDir = os.path.dirname(os.path.realpath(__file__))
         self.projectDir = os.path.dirname(self.moduleDir)
-        self.outputDir = os.path.join(self.projectDir, "mchartanalyzer-output")
+        self.outputDir = os.path.join(self.projectDir, "mchartanalyzer-data")
 
         self.testMode = True if testMode else False
