+++ mchartanalyzer/scrapestrategy.py
