+++ mchartanalyzer/ultimateguitarstrategy.py
